refactor(agent): extract run_one_tool + spawn_delta_forwarder into engine/
Phase 1 of unifying the three agent-turn loops (Agent::turn, run_tool_call_loop,
subagent run_inner_loop) behind one engine. Introduces
src/openhuman/agent/harness/engine/ and lifts two pieces that were duplicated
across the loops, behavior-preserving:

- engine::run_one_tool — the per-call tool executor (start event, policy gate,
  scope guard, approval gate, execute+timeout, scrub/tokenjuice/cap/summarize,
  audit, completion event). Collapses ~400 lines of branches in tool_loop.rs
  into one call + a uniform push/writeln/circuit-breaker tail.
- engine::spawn_delta_forwarder — the ProviderDelta -> AgentProgress streaming
  forwarder, previously copied byte-for-byte in tool_loop.rs and turn.rs.

No behavior change. tool_loop tests: 24/24 pass.
