+++ src/openhuman/agent/harness/mod.rs
@@ -22,6 +22,7 @@
 pub(crate) mod archivist;
 pub(crate) mod builtin_definitions;
 mod credentials;
+pub(crate) mod engine;
 pub mod definition;
 pub(crate) mod definition_loader;
 pub mod fork_context;
